requirements.txt
python-telegram-bot==13.5
requests==2.25
datetime==4.3
firebase-admin==5.0.0
